+++ bgameb/game.py
@@ -28,33 +28,41 @@ if __name__ == '__main__':
     game.game_steps.add(bgameb.Step('step 1', priority=1))
 
     game.add(bgameb.Deck('Cards Deck'))
-    card = bgameb.Card('One card')
-
+    card = bgameb.Card('One card', description='story')  # type: ignore
     game.cards_deck.add(card)
     game.cards_deck.one_card.count = 3
+
+    print(f'Other: {game.cards_deck.one_card.other}')
     game.cards_deck.one_card.counter['yellow'] = 12
     game.cards_deck.one_card.counter['banana'] = 0
 
     steps = game.game_steps.deal()
-    deck = game.cards_deck.deal()
+    game.cards_deck.deal()
+    deck = game.cards_deck.shuffle()
 
     game.IS_ACTIVE = True  # type: ignore
 
     game.add(bgameb.Shaker('blue shaker'))
-    game.blue_shaker.add(bgameb.Dice('dice#8', sides=8, count=10))
+    game.blue_shaker.add(
+        bgameb.Dice('dice#8', sides=8, count=10)
+            )
 
     result = game.blue_shaker.dice_8.roll()
 
-    print(result)
+    print(f'Result of shke: {result}')
     print('='*20 + '\n')
-    print(repr(game))
+    print(f'Repr: {repr(game)}')
     print('='*20 + '\n')
-    print(game)
+    print(f'Print: {game}')
     print('='*20 + '\n')
     pprint(game.to_dict())
     print('='*20 + '\n')
     pprint(deck)
     print('='*20 + '\n')
-    print(game.get_names())
+    print(game.cards_deck.get_current_names())
     print('='*20 + '\n')
-    print(dir(game))
+    print(f' Get_ids(): {game.get_names()}')
+    print('='*20 + '\n')
+    pprint(f'Len: {len(game)}')
+    print('='*20 + '\n')
+    print(f'Dir: {dir(game)}')